+++ cpp11test/src/test-as.cpp
@@ -2,6 +2,7 @@
 #include <string>
 #include <vector>
 
+#include "cpp11/R.hpp"
 #include "cpp11/declarations.hpp"
 
 #include <testthat.h>
@@ -370,7 +371,7 @@ context("as_cpp-C++") {
     x.push_back(2);
 
     auto res1 = cpp11::as_sexp(x);
-    expect_true(TYPEOF(res1) == RAWSXP);
+    expect_true(cpp11::detail::r_typeof(res1) == RAWSXP);
     expect_true(RAW(res1)[0] == 0);
     expect_true(RAW(res1)[1] == 1);
     expect_true(RAW(res1)[2] == 2);
@@ -378,7 +379,7 @@ context("as_cpp-C++") {
     cpp11::raws y(x);
 
     auto res2 = cpp11::as_sexp(y);
-    expect_true(TYPEOF(res2) == RAWSXP);
+    expect_true(cpp11::detail::r_typeof(res2) == RAWSXP);
     expect_true(RAW(res2)[0] == 0);
     expect_true(RAW(res2)[1] == 1);
     expect_true(RAW(res2)[2] == 2);